requirements.txt
pytesseract
opencv-python
Pillow
tkinterdnd2